Switches from register to login page on register button press -- all variables are in one file right now.
